:bug: Even with date range, get all runs for first seen
If we only call /first_seen for the requested time range then in particular the % new metric will not be correct as it will not take into account runs before the start date. That means some segments in runs during the date range will look as if they are new when in fact they are been run before, just not in the date range being shown.