:bug: Deal with invalid date inputs (#7)
Due to using date picker elements, an invalid date can only occur if a user types in one of the fields, and even then some possible failure modes can't occur, e.g. trying to type a day above 31 or month above 12 will not work. Only typing an invalid day for a shorter month, e.g. 31st November, will cause an error. In this case the date picker's value is just the empty string rather than what is visible in the date picker. The other invalid situation is when the start date is after the end date.