:bug: Fix bug with day diff after removing from API
Day difference was previously calculated in the backend between consecutive runs, hence when calculating the total delay it was additive. Switching to the calculation in the frontend is now doing so relative to the start date (which may be a date when we paused the animation). Thus the total delay is no longer additive since we computing the full delay for everything we want to put inside a setTimeout call.